Rename sample_module.py to sampler.py and resample.py to resampler.py
