+++ tests/unit/modules/test_resampler.py
