+++ tests/unit/modules/test_sampler.py
